requirements.txt
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.1
bleach==1.4.3
certifi==2015.11.20.1
chardet==2.3.0
colorlover==0.2.1
cufflinks==0.7.2
cycler==0.10.0
Cython==0.23.4
dateutils==0.6.6
decorator==4.0.9
Django==1.10
django-contrib-comments==1.7.2
djangorestframework==3.4.4
EasyProcess==0.2.2
feedparser==5.2.1
filebrowser-safe==0.4.5
functools32==3.2.3.post2
future==0.15.2
futures==3.0.5
googlemaps==2.4.3
grappelli-safe==0.4.4
gunicorn==19.6.0
html5lib==0.9999999
igraph==0.1.11
ipykernel==4.2.2
ipython==4.1.1
ipython-genutils==0.1.0
ipywidgets==4.1.1
jgraph==0.2.1
Jinja2==2.8
joblib==0.9.4
jsonschema==2.5.1
MarkupSafe==0.23
mistune==0.7.1
mlab==1.1.4
nbconvert==4.1.0
nbformat==4.0.1
networkx==1.11
notebook==4.1.0
oauthlib==1.0.3
path.py==8.1.2
pbr==1.10.0
pickleshare==0.6
ptyprocess==0.5.1
Pygments==2.1.1
pyparsing==2.1.0
pytesseract==0.1.6
python-dateutil==2.5.3
pytz==2016.6.1
PyVirtualDisplay==0.1.5
qtconsole==4.1.1
requests==2.9.1
requests-oauthlib==0.6.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
splinter==0.7.3
stevedore==1.17.0
terminado==0.6
tornado==4.3
traitlets==4.1.0
tweepy==3.5.0
tweetpy==0.1
twitter==1.17.1
tzlocal==1.2.2
urllib3==1.16
vincent==0.4.4
virtualenv==15.0.2
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.1
virtualenvwrapper-win==1.2.1
whitenoise==3.2.1
xmljson==0.1.6
xmltodict==0.10.2
